[Fix] メモリリークを解消。
git-svn-id: svn+ssh://macmini/usr/local/svnrepos/PreviewerSelector/PreviewerSelector@33 42e33f61-e417-4bac-a148-64f5df846d20

+++ PSPreference.m
@@ -174,7 +174,7 @@ enum _PreferenceMenuTags {
 
 - (NSDragOperation)tableView:(NSTableView*)targetTableView
 				validateDrop:(id <NSDraggingInfo>)info
-				 proposedRow:(int)row
+				 proposedRow:(NSInteger)row
 	   proposedDropOperation:(NSTableViewDropOperation)dropOperation
 {
 	NSPasteboard *pboard = [info draggingPasteboard];
@@ -196,7 +196,7 @@ enum _PreferenceMenuTags {
 
 - (BOOL)tableView:(NSTableView*)tableView
 	   acceptDrop:(id <NSDraggingInfo>)info
-			  row:(int)row
+			  row:(NSInteger)row
 	dropOperation:(NSTableViewDropOperation)dropOperation
 {
 	NSPasteboard *pboard = [info draggingPasteboard];
